Add JetBrains copyright notice to derived code
