+++ zipline-kotlin-plugin-hosted/src/main/kotlin/app/cash/zipline/kotlin/ir.kt
@@ -87,28 +87,6 @@ internal val IrSimpleFunction.signature: String
     append("fun ${name.identifier}(${valueParameters.joinToString { (it.type as IrSimpleType).asString() }}): ${(returnType as IrSimpleType).asString()}")
   }
 
-/** Inspired by [org.jetbrains.kotlin.ir.backend.js.utils.asString]. */
-fun IrSimpleType.asString(): String =
-  classifier.asString() +
-    (arguments.ifNotEmpty {
-      joinToString(separator = ",", prefix = "<", postfix = ">") { it.asString() }
-    } ?: "") +
-    (if (hasQuestionMark) "?" else "")
-
-/** Copied from [org.jetbrains.kotlin.ir.backend.js.utils.asString]. */
-private fun IrTypeArgument.asString(): String = when (this) {
-  is IrStarProjection -> "*"
-  is IrTypeProjection -> variance.label + (if (variance != Variance.INVARIANT) " " else "") + (type as IrSimpleType).asString()
-  else -> error("Unexpected kind of IrTypeArgument: " + javaClass.simpleName)
-}
-
-/** Copied from [org.jetbrains.kotlin.ir.backend.js.utils.asString]. */
-private fun IrClassifierSymbol.asString() = when (this) {
-  is IrTypeParameterSymbol -> this.owner.name.asString()
-  is IrClassSymbol -> this.owner.fqNameWhenAvailable!!.asString()
-  else -> error("Unexpected kind of IrClassifierSymbol: " + javaClass.typeName)
-}
-
 internal fun FqName.child(name: String) = child(Name.identifier(name))
 
 /** Thrown on invalid or unexpected input code. */
